Rename files to use underscore
